+++ Alert.php
@@ -135,9 +135,7 @@ class Alert extends Widget
      */
     protected function initOptions()
     {
-        Html::addCssClass($this->options, 'alert');
-        Html::addCssClass($this->options, 'fade');
-        Html::addCssClass($this->options, 'in');
+        Html::addCssClass($this->options, ['alert', 'fade', 'in']);
 
         if ($this->closeButton !== false) {
             $this->closeButton = array_merge([