+++ BaseHtml.php
@@ -20,8 +20,7 @@ use yii\helpers\ArrayHelper;
 class BaseHtml extends \yii\helpers\Html
 {
     /**
-     * Composes icon HTML.
-     * @see http://getbootstrap.com/components/#glyphicons
+     * Composes icon HTML for bootstrap Glyphicons.
      * @param string $name icon short name, for example: 'star'
      * @param array $options the tag options in terms of name-value pairs. These will be rendered as
      * the attributes of the resulting tag. There are also a special options:
@@ -30,6 +29,7 @@ class BaseHtml extends \yii\helpers\Html
      * - prefix: string, prefix which should be used to compose tag class, by default 'glyphicon glyphicon-' is used.
      *
      * @return string icon HTML.
+     * @see http://getbootstrap.com/components/#glyphicons
      */
     public static function icon($name, $options = [])
     {
@@ -41,7 +41,7 @@ class BaseHtml extends \yii\helpers\Html
 
     /**
      * Renders Bootstrap static form control.
-     * @see http://getbootstrap.com/css/#forms-controls-static
+     *
      * By default value will be HTML-encoded using [[encode()]], you may control this behavior
      * via 'encode' option.
      * @param string $value static control value.
@@ -51,6 +51,7 @@ class BaseHtml extends \yii\helpers\Html
      * - encode: boolean, whether value should be HTML-encoded or not.
      *
      * @return string generated HTML
+     * @see http://getbootstrap.com/css/#forms-controls-static
      */
     public static function staticControl($value, $options = [])
     {
@@ -72,6 +73,7 @@ class BaseHtml extends \yii\helpers\Html
      * about attribute expression.
      * @param array $options the tag options in terms of name-value pairs. See [[staticControl()]] for details.
      * @return string generated HTML
+     * @see staticControl()
      */
     public static function activeStaticControl($model, $attribute, $options = [])
     {
@@ -83,4 +85,4 @@ class BaseHtml extends \yii\helpers\Html
         }
         return static::staticControl($value, $options);
     }
-} 
+}