+++ Button.php
@@ -46,7 +46,7 @@ class Button extends Widget
     {
         parent::init();
         $this->clientOptions = false;
-        Html::addCssClass($this->options, 'btn');
+        Html::addCssClass($this->options, ['widget' => 'btn']);
     }
 
     /**