+++ NavBar.php
@@ -19,7 +19,7 @@ use yii\helpers\ArrayHelper;
  *
  * ```php
  * use yii\bootstrap\NavBar;
- * use yii\widgets\Menu;
+ * use yii\bootstrap\Nav;
  *
  * NavBar::begin(['brandLabel' => 'NavBar Test']);
  * echo Nav::widget([
@@ -27,6 +27,7 @@ use yii\helpers\ArrayHelper;
  *         ['label' => 'Home', 'url' => ['/site/index']],
  *         ['label' => 'About', 'url' => ['/site/about']],
  *     ],
+ *     'options' => ['class' => 'navbar-nav'],
  * ]);
  * NavBar::end();
  * ```
@@ -55,14 +56,15 @@ class NavBar extends Widget
      */
     public $containerOptions = [];
     /**
-     * @var string|boolean the text of the brand of false if it's not used. Note that this is not HTML-encoded.
+     * @var string|boolean the text of the brand or false if it's not used. Note that this is not HTML-encoded.
      * @see http://getbootstrap.com/components/#navbar
      */
     public $brandLabel = false;
     /**
-     * @param array|string|boolean $url the URL for the brand's hyperlink tag. This parameter will be processed by [[Url::to()]]
+     * @var array|string|boolean $url the URL for the brand's hyperlink tag. This parameter will be processed by [[\yii\helpers\Url::to()]]
      * and will be used for the "href" attribute of the brand link. Default value is false that means
      * [[\yii\web\Application::homeUrl]] will be used.
+     * You may set it to `null` if you want to have no link at all.
      */
     public $brandUrl = false;
     /**
@@ -93,11 +95,11 @@ class NavBar extends Widget
     {
         parent::init();
         $this->clientOptions = false;
-        Html::addCssClass($this->options, 'navbar');
-        if ($this->options['class'] === 'navbar') {
-            Html::addCssClass($this->options, 'navbar-default');
+        if (empty($this->options['class'])) {
+            Html::addCssClass($this->options, ['navbar', 'navbar-default']);
+        } else {
+            Html::addCssClass($this->options, ['widget' => 'navbar']);
         }
-        Html::addCssClass($this->brandOptions, 'navbar-brand');
         if (empty($this->options['role'])) {
             $this->options['role'] = 'navigation';
         }
@@ -116,12 +118,11 @@ class NavBar extends Widget
         }
         echo $this->renderToggleButton();
         if ($this->brandLabel !== false) {
-            Html::addCssClass($this->brandOptions, 'navbar-brand');
+            Html::addCssClass($this->brandOptions, ['widget' => 'navbar-brand']);
             echo Html::a($this->brandLabel, $this->brandUrl === false ? Yii::$app->homeUrl : $this->brandUrl, $this->brandOptions);
         }
         echo Html::endTag('div');
-        Html::addCssClass($this->containerOptions, 'collapse');
-        Html::addCssClass($this->containerOptions, 'navbar-collapse');
+        Html::addCssClass($this->containerOptions, ['collapse' => 'collapse', 'widget' => 'navbar-collapse']);
         $options = $this->containerOptions;
         $tag = ArrayHelper::remove($options, 'tag', 'div');
         echo Html::beginTag($tag, $options);