+++ Tabs.php
@@ -122,7 +122,7 @@ class Tabs extends Widget
     public function init()
     {
         parent::init();
-        Html::addCssClass($this->options, 'nav ' . $this->navType);
+        Html::addCssClass($this->options, ['widget' => 'nav', $this->navType]);
     }
 
     /**
@@ -162,21 +162,23 @@ class Tabs extends Widget
 
             if (isset($item['items'])) {
                 $label .= ' <b class="caret"></b>';
-                Html::addCssClass($headerOptions, 'dropdown');
+                Html::addCssClass($headerOptions, ['widget' => 'dropdown']);
 
                 if ($this->renderDropdown($n, $item['items'], $panes)) {
                     Html::addCssClass($headerOptions, 'active');
                 }
 
-                Html::addCssClass($linkOptions, 'dropdown-toggle');
-                $linkOptions['data-toggle'] = 'dropdown';
+                Html::addCssClass($linkOptions, ['widget' => 'dropdown-toggle']);
+                if (!isset($linkOptions['data-toggle'])) {
+                    $linkOptions['data-toggle'] = 'dropdown';
+                }
                 $header = Html::a($label, "#", $linkOptions) . "\n"
                     . Dropdown::widget(['items' => $item['items'], 'clientOptions' => false, 'view' => $this->getView()]);
             } else {
                 $options = array_merge($this->itemOptions, ArrayHelper::getValue($item, 'options', []));
                 $options['id'] = ArrayHelper::getValue($options, 'id', $this->options['id'] . '-tab' . $n);
 
-                Html::addCssClass($options, 'tab-pane');
+                Html::addCssClass($options, ['widget' => 'tab-pane']);
                 if (ArrayHelper::remove($item, 'active')) {
                     Html::addCssClass($options, 'active');
                     Html::addCssClass($headerOptions, 'active');
@@ -185,7 +187,9 @@ class Tabs extends Widget
                 if (isset($item['url'])) {
                     $header = Html::a($label, $item['url'], $linkOptions);
                 } else {
-                    $linkOptions['data-toggle'] = 'tab';
+                    if (!isset($linkOptions['data-toggle'])) {
+                        $linkOptions['data-toggle'] = 'tab';
+                    }
                     $header = Html::a($label, '#' . $options['id'], $linkOptions);
                 }
 
@@ -241,7 +245,7 @@ class Tabs extends Widget
 
             $content = ArrayHelper::remove($item, 'content');
             $options = ArrayHelper::remove($item, 'contentOptions', []);
-            Html::addCssClass($options, 'tab-pane');
+            Html::addCssClass($options, ['widget' => 'tab-pane']);
             if (ArrayHelper::remove($item, 'active')) {
                 Html::addCssClass($options, 'active');
                 Html::addCssClass($item['options'], 'active');
@@ -250,8 +254,9 @@ class Tabs extends Widget
 
             $options['id'] = ArrayHelper::getValue($options, 'id', $this->options['id'] . '-dd' . $itemNumber . '-tab' . $n);
             $item['url'] = '#' . $options['id'];
-            $item['linkOptions']['data-toggle'] = 'tab';
-
+            if (!isset($item['linkOptions']['data-toggle'])) {
+                $item['linkOptions']['data-toggle'] = 'tab';
+            }
             $panes[] = Html::tag('div', $content, $options);
 
             unset($item);