+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Yii Framework 2 bootstrap extension Change Log
 - Enh #43: Added support for the static form controls via `yii\bootstrap\Html` (klimov-paul)
 - Enh #44: Fixed `yii\bootstrap\ButtonDropdown` renders two buttons with the same id, if 'split' is enabled (klimov-paul)
 - Enh #50: Added `dropDownOptions` that is passed to `yii\bootstrap\Nav` dropdown items (fbau123)
+- Enh #92: Allow overriding `data-toggle` in `yii\bootstrap\Tabs` (machour)
 
 
 2.0.4 May 10, 2015