+++ CHANGELOG.md
@@ -1,16 +1,23 @@
 Yii Framework 2 bootstrap extension Change Log
 ==============================================
 
-2.0.5 under development
+2.0.6 under development
 -----------------------
 
+- Bug #98: Fixed `yii\bootstrap\ButtonDropdown` setting `href` attribute for non `a` tags (13nightevil)
+
+2.0.5 September 23, 2015
+------------------------
+
+- Enh #15: Allowed overriding default Bootstrap CSS classes added by widgets (klimov-paul)
 - Enh #38: Added object support for `content` option in `Collapse` class (pana1990, ItsReddi)
 - Enh #40: Added `visible` option to `yii\bootstrap\Tab` widget items (klimov-paul)
 - Enh #41: Added `submenuOptions` support at `yii\bootstrap\Dropdown` (spikyjt, klimov-paul)
 - Enh #42: Added support for the glyphicons via `yii\bootstrap\Html::icon()` (klimov-paul)
 - Enh #43: Added support for the static form controls via `yii\bootstrap\Html` (klimov-paul)
-- Enh #44: Fixed `yii\bootstrap\ButtonDropdown` renders two buttons have same with id if 'split' is enabled (klimov-paul)
-- Enh #45: Added support for Bootstrap checkbox/radio toggle buttons (RomeroMsk, klimov-paul)
+- Enh #44: Fixed `yii\bootstrap\ButtonDropdown` renders two buttons with the same id, if 'split' is enabled (klimov-paul)
+- Enh #50: Added `dropDownOptions` that is passed to `yii\bootstrap\Nav` dropdown items (fbau123)
+- Enh #92: Allow overriding `data-toggle` in `yii\bootstrap\Tabs` (machour)
 
 
 2.0.4 May 10, 2015